+++ procollab/settings.py
@@ -94,7 +94,7 @@ INSTALLED_APPS = [
     "partner_programs.apps.PartnerProgramsConfig",
     "mailing.apps.MailingConfig",
     "feed.apps.FeedConfig",
-    "rate_projects.apps.RateProjectsConfig",
+    "project_rates.apps.ProjectRatesConfig",
     # Rest framework
     "rest_framework",
     "rest_framework_simplejwt",
