+++ project_rates/constants.py
